+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## [0.0.1] - 09 Jan 2023.
 
-- initial implementation.
+- Initial implementation.
+
+## [0.0.2] - 10 Jan 2023.
+
+- Add Relays method.
+- Add GetRelay method.
+- Improve code quality.